==> 03_warp_other_masks.ipynb <==
# warp other masks to functional space

import glob
import os
import re
import shutil
import random
import string

import numpy as np
import pandas as pd

import nilearn
from nilearn import plotting, image
from nilearn.input_data import NiftiMasker
import nibabel as nib
from nipype.interfaces import ants
import nighres

import subprocess
import json
import multiprocessing as mp
from functools import partial
import joblib
from joblib import Parallel, delayed
import itertools
from utils import apply_warp

import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
all_subjects

## Warp hardvard oxford masks to functional space

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['024']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/MNI152NLin2009cAsym_1mm_label-(?P<label>\S+).nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/masks/HarvardOxford_MNI152NLin2009cAsym/*.nii.gz'))
    
    output_img_base = f'../derivatives/masks_cortex_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
#     print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']

        final_img_fn = output_img_base.format(label=mask_label)
        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(20) as p:
    p.map(warp_masks_2_t1w, all_subjects)

## warp harvard oxford striatum masks to t1 space
not sure which MNI152, but seem to overlap well with MNI152NLin2009cAsym_1mm

#all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['043', '044']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/MNI_1mm_label-(?P<label>\S+).nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/masks/HarvardOxford_striatum/*.nii.gz'))
    
    output_img_base = f'../derivatives/masks_striatum_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
#     print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']
        print(mask_label)
        final_img_fn = output_img_base.format(label=mask_label)
        print(final_img_fn)
        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(10) as p:
    p.map(warp_masks_2_t1w, all_subjects)

## warp harvard oxford striatum masks to t1 space
use MNI2009c folder

#all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['043', '044']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/MNI152NLin2009cAsym_1mm_label-(?P<label>\S+)_def-img.nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/masks/HarvardOxford_striatum_MNI152NLin2009cAsym/*.nii.gz'))
    
    output_img_base = f'../derivatives/masks_str_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
#     print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']
        print(mask_label)
        final_img_fn = output_img_base.format(label=mask_label)
        print(final_img_fn)
        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(10) as p:
    p.map(warp_masks_2_t1w, all_subjects)

# warp HCP_MMP1 IFG masks to functional space

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['024']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/HCP_MMP1_MNI152NLin2009cAsym_1mm_label-(?P<label>\S+).nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/masks/HCP_MMP1_MNI152NLin2009cAsym/*.nii.gz'))
    
    output_img_base = f'../derivatives/masks_HCP_MMP1_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
#     print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']

        final_img_fn = output_img_base.format(label=mask_label)
        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(20) as p:
    p.map(warp_masks_2_t1w, all_subjects)

# Convert AAL3 atlas from SPM MNI to MNI200cAsym

use ants registration in nighres to register weird SPM MNI image to MNI2009c

import nighres

def apply_ants(source,target):
    
    nighres.registration.embedded_antsreg(source,target,run_rigid=True,run_affine=False,ignore_header=True,ignore_affine=True, overwrite=True,
                                         save_data=True,output_dir='../sourcedata/masks/aal3_atlas',file_name='AAL3v1_MNI_to_MNI2009c.nii.gz')

source = '../sourcedata/masks/aal3_atlas/ROI_MNI_V7_1mm.nii.gz'
target = '../sourcedata/templates/mni_icbm152_t1_tal_nlin_asym_09c_brain.nii'
apply_ants(source,target)

import nipype
import nibabel as nib
import os
import random
import string
import glob

# make sure you dont input a 4D image
def apply_mni_warp(img,interpolation='Linear'):
    ''' img can be a file path (preferred) OR a nifti1Image (which will be temporarily saved and then provided as a filepath to ants) '''
    from nipype.interfaces import ants

    if isinstance(img, str):
        if not os.path.exists(img):
            raise(IOError('img must either be a string pointing to an existing file, or a Nifti1Image. You passed: {}'.format(img)))
        input_is_img = False
    elif isinstance(img, nib.Nifti1Image):
        # save file, use a random string in name
        import random
        import string

        tmp1_name = './tmp_img_to_warp-{}.nii.gz'.format(''.join(random.choices(string.ascii_letters + string.digits, k=10)))
        img.to_filename(tmp1_name)  # save to temporary file
        img = tmp1_name
        input_is_img = True
    else:
        raise(IOError('Input type not understood... img must either be a string pointing to an existing file, or a Nifti1Header. You passed: {}'.format(img)))
    
    # find affine & composite warp
#     if t1w_to_MNI:
    ## warping an img from functional T1w space (1.5 mm) to MNI2009c (1 mm)
    composite_warp_name = 'from-MNI_to-MNI152NLin2009b'
    template_brain = '../sourcedata/templates/mni_icbm152_nlin_sym_09b/mni_icbm152_t1_tal_nlin_sym_09b_hires.nii'
    out_postfix = '_warped_2_MNI2009b'
#     else:
#         ## warping an img from MNI2009c (1 mm) to functional T1w space (1.5 mm)
#         composite_warp_name = 'from-MNI152NLin2009cAsym_to-T1w'
#         # use first bold_ref as reference img
#         boldrefs = sorted(glob.glob(os.path.join('..', 'derivatives', 'fmriprep', 'fmriprep', f'sub-{sub}', 'ses-*', 'func', 
#                                       f'sub-*_ses-*_task-*_run-1_space-T1w_boldref.nii.gz')))
#         template_brain = boldrefs[0]
#         out_postfix = f'_warped_2_T1w_sub-{sub}'

    composite_xfm = '/home/Public/trondheim/sourcedata/templates/MNI_6thgen_2_MNI2009b.h5'
        
#     composite_xfm = os.path.join('..', 'derivatives', 'fmriprep', 'fmriprep', f'sub-{sub}', 'anat', 
#                              f'sub-{sub}_{composite_warp_name}_mode-image_xfm.h5')
#     if not os.path.exists(composite_xfm):
#         # if there's _only_ an anatomical session and no other sessions, the warp (.h5 file) is in a different place with a slightly different name...
#         composite_xfm = os.path.join('..', 'derivatives', 'fmriprep', 'fmriprep', f'sub-{sub}', 'ses-anatomical', 'anat', 
#                              f'sub-{sub}_ses-anatomical_{composite_warp_name}_mode-image_xfm.h5')
    
    output_img_name = os.path.basename(img).replace('.nii.gz', '') + out_postfix + '.nii.gz'
    
    # ants set-up
    warp = ants.ApplyTransforms()
    warp.inputs.input_image = img
    warp.inputs.input_image_type = 0
    warp.inputs.interpolation = interpolation
    warp.inputs.invert_transform_flags = [False] #,False]
    warp.inputs.reference_image = template_brain
    warp.inputs.out_postfix = out_postfix
    
#     if t1w_to_MNI:
#         warp.inputs.transforms = [composite_xfm]
#     else:
#     warp.inputs.transforms = [composite_xfm]
    
    res = warp.run()
    
    if input_is_img:
        os.remove(img)  # remove temporary file again, all we care is the output anyway
    return str(res.outputs.output_image)

new = apply_mni_warp('/home/Public/trondheim/sourcedata/masks/aal3_atlas/ROI_MNI_V7_1mm.nii')

!antsApplyTransforms --default-value 0 --float 0 --input /home/Public/trondheim/sourcedata/masks/aal3_atlas/ROI_MNI_V7_1mm.nii --input-image-type 0 --interpolation Linear --output ROI_MNI_V7_1mm_warped_2_MNI2009b.nii --reference-image ../sourcedata/templates/mni_icbm152_nlin_asym_09b/mni_icbm152_t1_tal_nlin_asym_09b_hires.nii --transform /home/Public/trondheim/sourcedata/templates/MNI_6thgen_2_MNI2009b.h5

## Pauli atlas



Pauli_labels = {
'1':  "Pu",
'2':   'Ca',
'3':   "NAC",
'4':   'EXA',
'5':   'GPe',
'6':   'GPi',
'7':   'SNc',
'8':   'RN',
'9':   'SNr', 
'10':   'PBP',
'11':  'VTA',
'12':  "VeP",
'13':  'HN',
'14':  'HTH',
'15':  'MN',
'16':  'STH',
}

print(Pauli_labels['3'])

### rename Pauli files in /home/Public/trondheim/sourcedata/masks/Pauli2017_atlas

structures_list =[x.split('/')[-1] for x in sorted(glob.glob('/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/Pauli_atlas/CIT168toMNI152_prob_atlas_bilat_1mm__*'))]
structures_list


# #rename Pauli files

# for x, i in enumerate(structures_list):

#         old_label = structures_list[x].split('.')[0]
#         print(old_label)
#         keep_label= old_label.split('_(')[0]
#         old_label= old_label.split('volume ')[1]
#         index = old_label.split(')')[0]
#         new_label = keep_label+Pauli_labels[index]
#         print(new_label)



import os


for x, i in enumerate(structures_list):

        original_label = structures_list[x].split('.')[0]
        print(original_label)
        keep_label= original_label.split('_(')[0]
        old_label= original_label.split('volume ')[1]
        index = old_label.split(')')[0]
        new_label = keep_label+Pauli_labels[index]
        print(new_label)

        old_file = os.path.join('/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/Pauli_atlas/', original_label+'.nii.gz')
        new_file = os.path.join('/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/Pauli_atlas/', new_label+'.nii.gz')
        os.rename(old_file, new_file)


### threshold and binarize

# masks =sorted(glob.glob('/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/Pauli_atlas/CIT168toMNI152_prob_atlas*'))
# masks

labels = [  "Pu",
  'Ca',
   "NAC",
   'EXA',
  'GPe',
  'GPi',
  'SNc',
   'RN',
   'SNr', 
   'PBP',
 'VTA',
 "VeP",
  'HN',
  'HTH',
 'MN',
 'STH']
labels

def threshold_and_binarize(labels):
    
    label = labels
    input_file = '/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/Pauli_atlas/CIT168toMNI152_prob_atlas_bilat_1mm_'+label+'.nii.gz'
    
    ouput_file_name = '/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/Pauli_atlas/bin_thresh30/CIT168toMNI152_prob_atlas_bilat_1mm_'+label+'_bin.nii.gz'
    
    
    command = 'fslmaths '+input_file+' -thr 0.3 -bin ' +ouput_file_name
    print(command)
    try:
         subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
    except subprocess.CalledProcessError as e:
        msg = 'execution failed (error code '+str(e.returncode)+')\n Output: '+str(e.output)
        raise subprocess.CalledProcessError(msg)

    
with mp.Pool(1) as p:
    p.map(threshold_and_binarize, labels)

    

### make bilateral masks / split masks


def split_bilateral_atlas(labels):
    
    label = labels
    input_file = '/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/Pauli_atlas/bin_thresh30/CIT168toMNI152_prob_atlas_bilat_1mm_'+label+'_bin.nii.gz'
    
    #ouput_file_name = 
    
    
    command = 'fslmaths '+input_file+' -roi 97 -1 -1 -1 -1 -1 -1 -1 /home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/'+label+'-r.nii.gz'
    print('create right '+label+ ' mask')
    print(command)
    try:
         subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
    except subprocess.CalledProcessError as e:
        msg = 'execution failed (error code '+str(e.returncode)+')\n Output: '+str(e.output)
        raise subprocess.CalledProcessError(msg)

    
    command = 'fslmaths '+input_file+' -roi -1 97 -1 -1 -1 -1 -1 -1 /home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/'+label+'-l.nii.gz'
    print('create left '+label+ ' mask')

    print(command)
    try:
         subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
    except subprocess.CalledProcessError as e:
        msg = 'execution failed (error code '+str(e.returncode)+')\n Output: '+str(e.output)
        raise subprocess.CalledProcessError(msg)
    
with mp.Pool(1) as p:
    p.map(split_bilateral_atlas, labels)

    

## warp Pauli atlas to t1

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
all_subjects

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['024']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/(?P<label>\S+).nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/masks/Pauli2017_atlas/*.nii.gz'))
    
    output_img_base = f'../derivatives/masks_Pauli_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
#     print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']

        final_img_fn = output_img_base.format(label=mask_label)
        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(20) as p:
    p.map(warp_masks_2_t1w, all_subjects)

## downsample to functional space

# already done in output of previous step somehow... so, no downsampling required?!

masks =[x.split('/')[-1].split('.')[0]  for x in sorted(glob.glob('/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/*.nii.gz'))]
masks

### requires all masks combined into one atlas... which is currently not the case for Pauli

or does warp_Pauli_map() do that?

## okay, I give up for now. does not work... 


def warp_Pauli_map(args, atlas, target_img):
    label_n, label = args
    this_map = atlas.get_fdata().copy()
    this_map[this_map!=(label_n+1)] = 0
    this_map[this_map>0] = 1
    this_map_nii = nib.Nifti1Image(this_map, atlas.affine)

    # ok now we need to do this rather annoying temporary save thing agian
    resampled_img = nilearn.image.resample_to_img(source_img=this_map_nii, target_img=target_img, interpolation='linear')
    return {label: resampled_img}


#
def load_Pauli_atlas_1p5mm(sub,overwrite=True): 
    labels = ['Ca-l','Ca-r', 'EXA-l', 'EXA-r', 'GPe-l', 'GPe-r', 'GPi-l', 'GPi-r', 'HN-l', 'HN-r',
              'HTH-l', 'HTH-r', 'MN-l', 'MN-r', 'NAC-l', 'NAC-r', 'PBP-l','PBP-r','Pu-l','Pu-r',
              'RN-l', 'RN-r', 'SNc-l', 'SNc-r', 'SNr-l', 'SNr-r', 'STH-l', 'STH-r', 'VTA-l', 'VTA-r','VeP-l','VeP-r']


    # save fns
    save_fn_template = f'/home/Public/trondheim/derivatives/masks_Pauli_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{}}.nii.gz'
    if os.path.exists(save_fn_template.format('Ca-l')) and not overwrite:
        # Already warped, load atlas
        maps = {}
        for label in labels:
            maps[label] = nib.load(save_fn_template.format(label))
        
    else:
        atlas_fn = os.path.join('home','Public','trondheim', 'derivatives', 'masks_Pauli_func', f'sub-{sub}', 'anat', f'sub-{sub}_space-T1w_desc-mask-{{}}.nii.gz')
        if not os.path.exists(atlas_fn):
            print(f'Pauli not yet available for sub-{sub} (returning 0 to prevent break)')
            return 0
        target_img = sorted(glob.glob(f'../derivatives/fmriprep/fmriprep/sub-{sub}/ses-*/func/sub-{sub}_ses-*_task-*_run-*_space-T1w_boldref.nii.gz'))[0]
        atlas = nib.load(atlas_fn)

        # now we need to reduce resolution to 1.5mm
        with mp.Pool(32) as p:
            maps = p.map(partial(warp_Pauli_map, atlas=atlas, target_img=target_img), enumerate(labels))
        maps = {list(x.keys())[0]: list(x.values())[0] for x in maps}
        
        # save
        os.makedirs(os.path.dirname(save_fn_template), exist_ok=True)
        for label in labels:
            maps[label].to_filename(save_fn_template.format(label))

    return maps

with mp.Pool(32) as p:
    maps = p.map(partial(warp_Pauli_map, atlas=atlas, target_img=target_img), enumerate(labels))
    maps = {list(x.keys())[0]: list(x.values())[0] for x in maps}
        
    # save
    os.makedirs(os.path.dirname(save_fn_template), exist_ok=True)
    for label in labels:
        maps[label].to_filename(save_fn_template.format(label))


all_subjects = [x.split('-')[1].split('.')[0] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub-*.html'))]
# exclude subs that aren't fmriprepped yet
# all_subjects = ['043','044']
for subject in all_subjects:
    print(subject)
    load_Pauli_atlas_1p5mm(subject)
    

def find_rois(sub, atlas_name='ATAG', space='T1w'):
    if atlas_name == 'ATAG':
        if space == 'MNI152NLin2009cAsym' or space == 'mni':
            ### Rois in MNI09c-space
            mask_dir='/home/Public/trondheim/sourcedata/masks/MNI152NLin2009cAsym_res-1p5'
            fns = sorted(glob.glob(mask_dir + '/space-*'))
            names = [re.match('.*space-(?P<space>[a-zA-Z0-9]+)_res-1p5_label-(?P<label>[a-zA-Z0-9]+)_probseg_def-img.nii.gz', fn).groupdict()['label'] for fn in fns]
        else:
            mask_dir = f'../derivatives/masks_atag_func/sub-{sub}/anat/sub-{sub}_*.nii.gz'
            fns = sorted(glob.glob(mask_dir))
            names = [re.match('.*space-(?P<space>[a-zA-Z0-9]+)_desc-mask-(?P<label>[a-zA-Z0-9]+).nii.gz', fn).groupdict()['label'] for fn in fns]
    elif atlas_name == 'MASSP':
        mask_dir = f'../derivatives/masks_massp_func/sub-{sub}/anat/sub-{sub}_*.nii.gz'
        fns = sorted(glob.glob(mask_dir))
        names = [re.match('.*space-(?P<space>[a-zA-Z0-9]+)_desc-mask-(?P<label>\S+).nii.gz', fn).groupdict()['label'] for fn in fns]
    elif atlas_name == 'Pauli':
        mask_dir = f'../derivatives/masks_Pauli_func/sub-{sub}/anat/sub-{sub}_*.nii.gz'
        fns = sorted(glob.glob(mask_dir))
        names = [re.match('.*space-(?P<space>[a-zA-Z0-9]+)_desc-mask-(?P<label>\S+).nii.gz', fn).groupdict()['label'] for fn in fns]

    roi_dict = dict(zip(names, fns))
    return roi_dict

def load_atlas(sub, atlas_name='Pauli', space='T1w'):
    from nilearn import image
    
    roi_dict = find_rois(sub, atlas_name, space)
    combined = image.concat_imgs(roi_dict.values())
    
    class AttrDict(dict):
        def __init__(self, *args, **kwargs):
            super(AttrDict, self).__init__(*args, **kwargs)
            self.__dict__ = self
            
    roi_atlas = AttrDict({'maps': combined,
                          'labels': roi_dict.keys()})
    
    return roi_atlas

masks_t1w = nib.load('../derivatives/masks_Pauli_func/sub-004/anat/sub-004_space-T1w_desc-mask-Ca-l.nii.gz')
t1w = nib.load('../derivatives/fmriprep/fmriprep/sub-004/anat/sub-004_desc-preproc_T1w.nii.gz')
plotting.plot_stat_map(masks_t1w, bg_img=t1w)

### looks weird here but looks fine in fseleyes

sub = '004'
boldref = sorted(glob.glob(f'../derivatives/fmriprep/fmriprep/sub-{sub}/ses-*/func/sub-{sub}_ses-*_task-*_run-*_space-T1w_boldref.nii.gz'))[3]
nilearn.plotting.plot_prob_atlas(load_atlas(sub).maps, bg_img=boldref)

### ignore below. The lead_dbs atlas version was fucked up, so downloaded it again from neurovault which is already in MNI09c asym space

### coregister Pauli template to mni_icbm152_t1_tal_nlin_asym_09c_brain.nii

rigid registration failed (brain shifted but affines matched)

trying affine registration

import nighres

def apply_ants(source,target):
    
    nighres.registration.embedded_antsreg(source,target,run_rigid=False,run_affine=True,ignore_header=True,ignore_affine=True, overwrite=True,
                                         save_data=True,output_dir='../sourcedata/masks/Pauli2017_atlas/template',file_name='Pauli2017_MNItemplate_to_MNI09c_asym.nii.gz')

    
source = '/home/atrutti1/Documents/CIT168_T1w_700um.nii.gz'
target = '../sourcedata/templates/mni_icbm152_t1_tal_nlin_asym_09c_brain.nii'
apply_ants(source,target)


structures =sorted(glob.glob('/home/Public/trondheim/sourcedata/lead_dbs_atlases/CIT168_Reinf_Learn*/mixed/*.nii.gz'))
structures
#structures = [x.split('/')[-1].split('-')[-1] for x in sorted(glob.glob('/home/Public/trondheim/sourcedata/lead_dbs_atlases/CIT168_Reinf_Learn*/mixed/*.nii.gz'))]



# the Pauli at;as affines do not seem to match the template...

import nighres

def apply_ants(source,target):
    
    nighres.registration.embedded_antsreg(source,target,run_rigid=False,run_affine=True,ignore_header=True,ignore_affine=True, overwrite=True,
                                         save_data=True,output_dir='../sourcedata/masks/Pauli2017_atlas/template',file_name='Pauli2017_MNItemplate_to_MNI09c_asym.nii.gz')

    
source = '/home/atrutti1/Documents/CIT168_T1w_700um.nii.gz'
target = '../sourcedata/templates/mni_icbm152_t1_tal_nlin_asym_09c_brain.nii'
apply_ants(source,target)


structures_list = [x.split('/')[-1] for x in sorted(glob.glob('/home/Public/trondheim/sourcedata/lead_dbs_atlases/CIT168_Reinf_Learn*/mixed/*.nii.gz'))]
structures_list
#structures = [x.split('/')[-1].split('-')[-1] for x in sorted(glob.glob('/home/Public/trondheim/sourcedata/lead_dbs_atlases/CIT168_Reinf_Learn*/mixed/*.nii.gz'))]



structures = []
for x, i in enumerate(structures_list):

    label = structures_list[x].split('.')[0]
    
    structures.append(label)
structures

### transform Pauli masks to mni_icbm152_t1_tal_nlin_asym_09c_brain.nii space


def transform_atlas2mni(structure):
    
    print(structure)
    image = '/home/Public/trondheim/sourcedata/lead_dbs_atlases/CIT168_Reinf_Learn (Pauli 2017) - imported from CIT168//mixed/'+structure+'.nii.gz'
    mapping = '/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/template/Pauli2017_MNItemplate_to_MNI09c_asym_ants-map.nii.gz'
    output_dir='/home/Public/trondheim/sourcedata/masks/Pauli2017_atlas/'
    file_name = structure+'_MNI9c_asym_'
    #print(file_name)
    nighres.registration.apply_coordinate_mappings(image, mapping, output_dir = output_dir, file_name = file_name, save_data=True)
    
with mp.Pool(1) as p:
    p.map(transform_atlas2mni, structures) 
    

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['024']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/MNI152NLin2009cAsym_1mm_label-(?P<label>\S+).nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/lead_dbs_atlases/CIT168_Reinf_Learn*/*.nii.gz'))
    
    output_img_base = f'../derivatives/masks_Pauli2017_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
#     print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']

        final_img_fn = output_img_base.format(label=mask_label)
        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(20) as p:
    p.map(warp_masks_2_t1w, all_subjects)

# warp Pijnenburg Brodmann atlas

## co-register MNI templates


import nighres

def apply_ants(source,target):
    
    nighres.registration.embedded_antsreg(source,target,run_rigid=True,run_affine=False,ignore_header=True,ignore_affine=True, overwrite=True,
                                         save_data=True,output_dir='../sourcedata/masks/Pijnenburg_Brodmann',file_name='MNI09a-nlin-sym_to_MNI2009c.nii.gz')

source = '../sourcedata/templates/mni_icbm152_nlin_sym_09a/mni_icbm152_t1_tal_nlin_sym_09a_brain.nii.gz'
target = '../sourcedata/templates/mni_icbm152_t1_tal_nlin_asym_09c_brain.nii'
apply_ants(source,target)

### check affines

# good!

### convert atlas to MNI09c asym template

import nighres

def apply_ants(source,target):
    
    nighres.registration.embedded_antsreg(source,target,run_rigid=True,run_affine=False,ignore_header=True,ignore_affine=True, overwrite=True,
                                         save_data=True,output_dir='../sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/',file_name='BrodmannICBM152_2_MNI2009c.nii.gz')

source = '../sourcedata/masks/Pijnenburg_Brodmann/Brodmann/Brodmann_ICBM152.nii.gz'
target = '../sourcedata/templates/mni_icbm152_t1_tal_nlin_asym_09c_brain.nii'
apply_ants(source,target)

### extract Brodmann areas to make dlPFC, mPFC and PPC masks

extract_lh_structures= ['fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 1007 -uthr 1007 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/dlPFC_BA8-BA9-BA46/BA8-l.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 1008 -uthr 1008 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/dlPFC_BA8-BA9-BA46/BA9-l.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 1038 -uthr 1038 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/dlPFC_BA8-BA9-BA46/BA46-l.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 1020 -uthr 1020 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/mPFC_BA24_BA32/BA24-l.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 1026 -uthr 1026 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/mPFC_BA24_BA32/BA32-l.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 1006 -uthr 1006 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/PPC_BA7_BA40/BA74-l.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 1033 -uthr 1033 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/PPC_BA7_BA40/BA40-l.nii.gz']
                        
for command in extract_lh_structures:           

    print(command)
    try:
        subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
    except subprocess.CalledProcessError as e:
        msg = 'execution failed (error code '+str(e.returncode)+')\n Output: '+str(e.output)
        raise subprocess.CalledProcessError(msg)

                         


extract_rh_structures= ['fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 2007 -uthr 2007 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/dlPFC_BA8-BA9-BA46/BA8-r.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 2008 -uthr 2008 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/dlPFC_BA8-BA9-BA46/BA9-r.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 2038 -uthr 2038 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/dlPFC_BA8-BA9-BA46/BA46-r.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 2020 -uthr 2020 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/mPFC_BA24_BA32/BA24-r.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 2026 -uthr 2026 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/mPFC_BA24_BA32/BA32-r.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 2006 -uthr 2006 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/PPC_BA7_BA40/BA74-r.nii.gz',
                        'fslmaths /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/BrodmannICBM152_2_MNI2009c_ants-def0.nii.gz -thr 2033 -uthr 2033 /home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/PPC_BA7_BA40/BA40-r.nii.gz']
                        
for command in extract_rh_structures:           

    print(command)
    try:
        subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
    except subprocess.CalledProcessError as e:
        msg = 'execution failed (error code '+str(e.returncode)+')\n Output: '+str(e.output)
        raise subprocess.CalledProcessError(msg)

                         


### combine extracted Brodmann areas to make dlPFC, mPFC and PPC masks

masks = [x.split('/')[-1].split('_')[0] for x in sorted(glob.glob('/home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/*BA*'))]
masks

directories = sorted(glob.glob('/home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/*BA*'))
directories

hemispheres = ['l', 'r']


# for directory in directories: 
for mask in masks:
    directory = '/home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/'+mask+'_*'
    output_dir = '/home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/FPN_masks'
#     print(directory)
    for hem in hemispheres:
        list_structures = sorted(glob.glob(directory+'/*'+hem+'.nii.gz'))
#         print(list_structures)
        #if len(list_structures) > 2:

        if mask == 'dlPFC':
            command = 'fslmaths '+list_structures[0]+ ' -add '+list_structures[1]+ ' -add '+list_structures[2]+ ' '+output_dir+'/'+mask+'-'+hem+'.nii.gz'
        else:
            command = 'fslmaths '+list_structures[0]+ ' -add '+list_structures[1]+' '+output_dir+'/'+mask+'-'+hem+'.nii.gz'

        print(command)
        try:
            subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
        except subprocess.CalledProcessError as e:
            msg = 'execution failed (error code '+str(e.returncode)+')\n Output: '+str(e.output)
            raise subprocess.CalledProcessError(msg)


### binarize


FPN_masks = glob.glob('/home/Public/trondheim/sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/FPN_masks/*')

for mask in FPN_masks:
#     print(mask)
    
    command = 'fslmaths '+mask+' -bin '+mask
    
    print(command)
    try:
        subprocess.check_output(command, shell=True, stderr=subprocess.STDOUT)
    except subprocess.CalledProcessError as e:
        msg = 'execution failed (error code '+str(e.returncode)+')\n Output: '+str(e.output)
        raise subprocess.CalledProcessError(msg)



### warp to t1

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['024']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/(?P<label>\S+).nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/FPN_masks/*.nii.gz'))
#     print(all_masks)
    output_img_base = f'../derivatives/masks_FPN_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
    print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']
#         print(mask_label)
        final_img_fn = output_img_base.format(label=mask_label)
        print(final_img_fn)

        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(20) as p:
    p.map(warp_masks_2_t1w, all_subjects)

# select and warp masks for Nir-Cohen replication

1. FPN masks

        dlPFC --> BA 8, BA 9, BA 46
         mPFC\ACC --> BA 24, BA 32(we also have an ACC mask though...)
         PPC --> BA 7, BA 40
        --> FPN_masks: ../derivatives/masks_FPN_func/

2. subcortical masks

         caudate -->  Pauli Ca 
         putamen -->  Pauli Pu
        --> derivatives/masks_Pauli_func/

         pallidum --> MASSP GPi, GPe
         thalamus --> MASSP Tha
        --> derivatives/masks_massp_func/



for future analyses (to also look at midbrain) also: 


     MASSP:
        SN
        VTA
        PAG
        STN
        PPN
        RN

     Pauli:
        SNc
        SNr
        VTA
        PBP

### 1. warp FPN masks

all_subjects = [x.split('/')[-2].split('-')[-1] for x in sorted(glob.glob('../derivatives/fmriprep/fmriprep/sub*/ses-anatomical'))]
# all_subjects = ['024']
def warp_masks_2_t1w(sub, overwrite=True):
    print(sub)
    regex_masks = re.compile('.*/(?P<label>\S+).nii.gz')
    all_masks = sorted(glob.glob('../sourcedata/masks/Pijnenburg_Brodmann/MNI09c_Brodmann/FPN_masks/*.nii.gz'))
#     print(all_masks)
    output_img_base = f'../derivatives/masks_WM-rep_func/sub-{sub}/anat/sub-{sub}_space-T1w_desc-mask-{{label}}.nii.gz'
    print(output_img_base)
    for mask in all_masks:
        mask_label = regex_masks.match(mask).groupdict()['label']
#         print(mask_label)
        final_img_fn = output_img_base.format(label=mask_label)
        print(final_img_fn)

        if not os.path.exists(final_img_fn) or overwrite:
            os.makedirs(os.path.dirname(final_img_fn), exist_ok=True)
            tmp_img_name = apply_warp(mask, sub=sub, t1w_to_MNI=False)
            shutil.move(tmp_img_name, final_img_fn)
            
# _ = [warp_masks_2_t1w(x) for x in all_subjects]
with mp.Pool(20) as p:
    p.map(warp_masks_2_t1w, all_subjects)

### 2. copy Pauli masks

--> Caudate, Putamen, accumbens, VTA, PBP, SNr, SNc

--> also rename VTA to VTAnc

    cd /home/Public/trondheim/derivatives/masks_Pauli_func


    for subject in $(ls -d *)
    do
    echo ${subject}
    cd ${subject}/anat/
    scp ${subject}_space-T1w_desc-mask-Pu-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-Pu-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-Pu-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-Pu-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-Ca-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-Ca-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-Ca-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-Ca-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-NAC-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-NAC-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-NAC-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-NAC-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-PBP-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-PBP-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-PBP-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-PBP-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-VTA-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-VTAnc-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-VTA-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-VTAnc-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-SNr-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SNr-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-SNr-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SNr-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-SNr-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SNr-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-SNr-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SNr-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-SNc-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SNc-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-SNc-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SNc-l.nii.gz
    cd ../../
    done

### 3. copy massp masks

        SN
        VTA
        PAG
        STN
        PPN
        RN

        GPi
        GPe
        Tha

    cd /home/Public/trondheim/derivatives/masks_massp_func
#$(ls -d *)

    for subject in $(ls -d sub-026) 
    do
    echo ${subject}
    cd ${subject}/anat/
    scp ${subject}_space-T1w_desc-mask-VTA-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-VTA-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-VTA-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-VTA-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-Tha-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-Tha-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-Tha-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-Tha-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-SN-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SN-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-SN-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-SN-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-GPi-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-GPi-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-GPi-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-GPi-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-GPe-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-GPe-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-GPe-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-GPe-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-PAG-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-PAG-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-PAG-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-PAG-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-PPN-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-PPN-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-PPN-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-PPN-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-STN-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-STN-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-STN-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-STN-r.nii.gz
    scp ${subject}_space-T1w_desc-mask-RN-l.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-RN-l.nii.gz
    scp ${subject}_space-T1w_desc-mask-RN-r.nii.gz /home/Public/trondheim/derivatives/masks_WM-rep_func/${subject}/anat/${subject}_space-T1w_desc-mask-RN-r.nii.gz
    cd ../../
    done
